# backend/src/docs/swagger.yaml
basePath: /v1
definitions:
...
  model.Address:
    properties:
      address:
        type: string
      apartment:
        type: string
      country:
        type: string
      createdAt:
        type: string
      district:
        type: string
      fullName:
        type: string
      id:
        type: string
      isDefault:
        type: boolean
      phone:
        type: string
      province:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
      ward:
        type: string
    type: object
  model.CartItem:
    properties:
      createdAt:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      quantity:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.Category:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  model.Image:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isPrimary:
        type: boolean
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      url:
        type: string
    type: object
  model.Order:
    properties:
      createdAt:
        type: string
      id:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      paymentMethod:
        $ref: '#/definitions/model.PaymentMethod'
      shippingAddress:
        type: string
      status:
        $ref: '#/definitions/model.OrderStatus'
      total:
        type: number
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.OrderItem:
    properties:
      id:
        type: string
      order:
        $ref: '#/definitions/model.Order'
      orderId:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      quantity:
        type: integer
    type: object
  model.OrderStatus:
    enum:
    - PENDING
    - PAID
    - SHIPPED
    - DELIVERED
    - CANCELLED
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusPaid
    - OrderStatusShipped
    - OrderStatusDelivered
    - OrderStatusCancelled
  model.PaymentMethod:
    enum:
    - COD
    - BANK_TRANSFER
    - STRIPE
    - MOMO
    - VNPAY
    type: string
    x-enum-varnames:
    - PaymentMethodCOD
    - PaymentMethodBankTransfer
    - PaymentMethodStripe
    - PaymentMethodMomo
    - PaymentMethodVnpay
  model.Product:
    properties:
      cartItems:
        items:
          $ref: '#/definitions/model.CartItem'
        type: array
      category:
        $ref: '#/definitions/model.Category'
      categoryId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      discount:
        type: number
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      name:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      price:
        type: number
      quantity:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
      wishlist:
        items:
          $ref: '#/definitions/model.Wishlist'
        type: array
    type: object
  model.Review:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      rating:
        type: integer
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.User:
    properties:
      address:
        type: string
      addresses:
        items:
          $ref: '#/definitions/model.Address'
        type: array
      bio:
        type: string
      cartItems:
        items:
          $ref: '#/definitions/model.CartItem'
        type: array
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isAdmin:
        type: boolean
      lastLoginAt:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
      phone:
        type: string
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      updatedAt:
        type: string
      wishlist:
        items:
          $ref: '#/definitions/model.Wishlist'
        type: array
    type: object
  model.Wishlist:
    properties:
      createdAt:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@birdnest.vn
    name: API Support
  description: BirdNest Shop REST API for e-commerce platform
  license:
    name: MIT
    url: https://github.com/indrayyana/go-fiber-boilerplate/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: BirdNest Shop API
  version: 1.0.0
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category (Admin only)
      tags:
      - Categories
...
  /admin/products:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product (Admin only)
      tags:
      - Products
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product (Admin only)
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing product (Admin only)
      tags:
      - Products
...
  /categories:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all categories
      tags:
      - Categories
...
  /products:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Category ID filter
        in: query
        name: categoryId
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Minimum price
        in: query
        name: minPrice
        type: number
      - description: Maximum price
        in: query
        name: maxPrice
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Products with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products with filtering and pagination
      tags:
      - Products
  /products/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a single product by ID
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"